third_party/WebKit/LayoutTests/fast/css/webkit-column-span-legacy.html

<!DOCTYPE HTML>
<script src="../../resources/js-test.js"></script>

<div id="-webkit-column-span" class="test" style="-webkit-column-span: 1;"></div>

<script>
var div1 = document.getElementById("-webkit-column-span");
shouldBe("getComputedStyle(div1).webkitColumnSpan", "'none'");
</script>
