third_party/blink/web_tests/platform/generic/fast/css/webkit-column-span-legacy-expected.txt

PASS getComputedStyle(div1).webkitColumnSpan is 'none'
PASS successfullyParsed is true

TEST COMPLETE
